youtube_analytics_project/requirements.txt
streamlit
google-api-python-client
pandas
matplotlib
cassandra-driver
langflow